# national-broker/src/main/openapi/DSDOrganizationApi.json
{
  "openapi" : "3.0.99",
  "info" : {
    "title" : "DSD Message service handler API",
    "description" : "This is a pilot project for  implementing ISA2 messaging REST API",
    "termsOfService" : "https://www.eupl.eu/",
    "contact" : {
      "name" : "Test user"
    },
    "license" : {
      "name" : "EUPL 1.2",
      "url" : "https://www.eupl.eu/"
    },
    "version" : "v1.0",
    "summary" : "DSD messaging rest service",
    "x-edel-lifecycle" : {
      "maturity" : "development"
    },
    "x-edel-publisher" : {
      "name" : "European Commission",
      "url" : "https://joinup.ec.europa.eu/collection/api4dt"
    }
  },
...
  "servers" : [ {
    "url" : "http://localhost:8080/dsd-mock",
    "description" : "Test server!"
  } ],
...
  "paths" : {
...
    "/v1/messaging/organization/update/{messageId}/response" : {
      "get" : {
        "tags" : [ "DSD Organization: Response Message Reference List" ],
        "summary" : "Get Response Message Reference List Endpoint",
        "description" : "This endpoint returns a list of response message references available for pulling, following the Message Reference schema, representing responses to a previous message sent by the original sender",
        "operationId" : "GetOrganizationResponseMessageReferenceListId",
        "parameters" : [ {
          "$ref" : "#/components/parameters/messageId"
        } ],
        "responses" : {
          "202" : {
            "description" : "Message reference list",
            "headers" : {
              "Original-Sender" : {
                "$ref" : "#/components/headers/Original-Sender"
              },
              "Final-Recipient" : {
                "$ref" : "#/components/headers/Final-Recipient"
              },
              "Timestamp" : {
                "$ref" : "#/components/headers/Timestamp"
              },
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageReferenceList"
                },
                "example" : {
                  "messageReferenceList" : [ {
                    "messageId" : "7199e0ef-fe0f-403e-9672-018a930709a3",
                    "service" : "my-service",
                    "action" : "my-action",
                    "href" : "/my-service/my-action/7199e0ef-fe0f-403e-9672-018a930709a3"
                  }, {
                    "messageId" : "03d2a204-47fe-431c-a20d-fbeab220e735",
                    "service" : "my-service",
                    "action" : "my-action",
                    "href" : "/my-service/my-action/03d2a204-47fe-431c-a20d-fbeab220e735"
                  } ],
                  "limit" : -1,
                  "offset" : 0
                }
              }
            }
          },
          "400" : {
            "description" : "Sent when the message fails the validation process",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "ValidationFailed" : {
                    "value" : {
                      "title" : "Validation Failed",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#message-validation-failed",
                      "status" : 400,
                      "detail" : "Sent when the message fails the validation process"
                    }
                  },
                  "InvalidMessageId" : {
                    "value" : {
                      "title" : "Invalid or Duplicate Message ID",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-message-id",
                      "status" : 400,
                      "detail" : "Sent when the MessageId is not valid"
                    }
                  },
                  "InvalidSignature" : {
                    "value" : {
                      "title" : "Invalid Message Signature",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-message-signature",
                      "status" : 400,
                      "detail" : "Sent when the message signature cannot be verified"
                    }
                  },
                  "InvalidAddressing" : {
                    "value" : {
                      "title" : "Invalid Addressing",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-addressing",
                      "status" : 400,
                      "detail" : "Sent when the Original Sender or Final Recipient(s) cannot be resolved"
                    }
                  },
                  "InvalidMessageFormat" : {
                    "value" : {
                      "title" : "Invalid Message Format",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-format",
                      "status" : 400,
                      "detail" : "Sent when the message format does not adhere to the specification"
                    }
                  },
                  "NoFinalRecipient" : {
                    "value" : {
                      "title" : "No final recipient configured for the pulling user",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#pull/no-final-recipient",
                      "status" : 400,
                      "detail" : "Sent when the server cannot resolve/match the pulling user to a final recipient"
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Sent when no message is found that maps to the pull request",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "NoMessageFound" : {
                    "value" : {
                      "title" : "No Message Found",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#pull/no-message-found",
                      "status" : 404,
                      "detail" : "Sent when no message is found that maps to the pull request"
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Sent when the pull request is unauthorized",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "Unauthorized" : {
                    "value" : {
                      "title" : "Unauthorized",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#pull/unauthorized",
                      "status" : 401,
                      "detail" : "Sent when the pull request is unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "InternalServerError" : {
                    "value" : {
                      "title" : "Internal Server Error",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md/internal-server-error",
                      "status" : 500,
                      "detail" : "The server encountered an unexpected condition that prevented it from fulfilling the request"
                    }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "DSD_ClientCredentials_OAuthSecurity" : [ "ROLE_DSD" ]
        } ]
      }
    },
...
    "/v1/messaging/{messageId}/response/signal" : {
      "post" : {
        "tags" : [ "DSD Organization: Message submission" ],
        "summary" : "Organization Signal Submission Endpoint ",
        "description" : "The Organization Signal Submission endpoint is used for signalling the successfully pulled message or to response errors/warnings at receiving messages.",
        "operationId" : "OrganizationSignalSubmissionId",
        "parameters" : [ {
          "$ref" : "#/components/parameters/Original-Sender"
        }, {
          "$ref" : "#/components/parameters/Final-Recipient"
        }, {
          "$ref" : "#/components/parameters/Timestamp"
        }, {
          "$ref" : "#/components/parameters/Edel-Message-Sig"
        }, {
          "$ref" : "#/components/parameters/messageId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignalMessage"
              },
              "example" : {
                "title" : "Message Accepted",
                "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#message-accepted",
                "status" : 202,
                "detail" : "Sent when the message is properly validated. It may include a status monitor that can provide the user with an estimate of when the request will be fulfilled (see [RFC7231])",
                "instance" : "/my-service/my-action/dde12f67-c391-4851-8fa2-c07dd8532efd",
                "digest" : "sha-256=eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "Sent when the message is properly validated and received.",
            "headers" : {
              "Original-Sender" : {
                "$ref" : "#/components/headers/Original-Sender"
              },
              "Final-Recipient" : {
                "$ref" : "#/components/headers/Final-Recipient"
              },
              "Timestamp" : {
                "$ref" : "#/components/headers/Timestamp"
              },
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignalMessage"
                },
                "example" : {
                  "title" : "Message Accepted",
                  "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#message-accepted",
                  "status" : 202,
                  "detail" : "Sent when the message is properly validated. It may include a status monitor that can provide the user with an estimate of when the request will be fulfilled (see [RFC7231])",
                  "instance" : "/my-service/my-action/dde12f67-c391-4851-8fa2-c07dd8532efd",
                  "digest" : "sha-256=eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ"
                }
              }
            }
          },
          "400" : {
            "description" : "Sent when the message fails the validation process",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "ValidationFailed" : {
                    "value" : {
                      "title" : "Validation Failed",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#message-validation-failed",
                      "status" : 400,
                      "detail" : "Sent when the message fails the validation process"
                    }
                  },
                  "InvalidMessageId" : {
                    "value" : {
                      "title" : "Invalid or Duplicate Message ID",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-message-id",
                      "status" : 400,
                      "detail" : "Sent when the MessageId is not valid"
                    }
                  },
                  "InvalidSignature" : {
                    "value" : {
                      "title" : "Invalid Message Signature",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-message-signature",
                      "status" : 400,
                      "detail" : "Sent when the message signature cannot be verified"
                    }
                  },
                  "InvalidAddressing" : {
                    "value" : {
                      "title" : "Invalid Addressing",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-addressing",
                      "status" : 400,
                      "detail" : "Sent when the Original Sender or Final Recipient(s) cannot be resolved"
                    }
                  },
                  "InvalidMessageFormat" : {
                    "value" : {
                      "title" : "Invalid Message Format",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#invalid-format",
                      "status" : 400,
                      "detail" : "Sent when the message format does not adhere to the specification"
                    }
                  },
                  "NoFinalRecipient" : {
                    "value" : {
                      "title" : "No final recipient configured for the pulling user",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#pull/no-final-recipient",
                      "status" : 400,
                      "detail" : "Sent when the server cannot resolve/match the pulling user to a final recipient"
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Sent when no message is found that maps to the pull request",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "NoMessageFound" : {
                    "value" : {
                      "title" : "No Message Found",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#pull/no-message-found",
                      "status" : 404,
                      "detail" : "Sent when no message is found that maps to the pull request"
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Sent when the pull request is unauthorized",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "Unauthorized" : {
                    "value" : {
                      "title" : "Unauthorized",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md#pull/unauthorized",
                      "status" : 401,
                      "detail" : "Sent when the pull request is unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request",
            "headers" : {
              "Edel-Message-Sig" : {
                "$ref" : "#/components/headers/Edel-Message-Sig"
              }
            },
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                },
                "examples" : {
                  "InternalServerError" : {
                    "value" : {
                      "title" : "Internal Server Error",
                      "type" : "https://github.com/isa2-api4ips/rest-api-profile/blob/main/messaging-api-specification/signal.md/internal-server-error",
                      "status" : 500,
                      "detail" : "The server encountered an unexpected condition that prevented it from fulfilling the request"
                    }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "DSD_ClientCredentials_OAuthSecurity" : [ "ROLE_DSD" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "MessageReferenceList" : {
        "title" : "A Message Reference object defined by the ISA² IPS REST API Messaging API Specification",
        "required" : [ "messageReferenceList" ],
        "type" : "object",
        "properties" : {
          "messageReferenceList" : {
            "type" : "array",
            "description" : "List of message references.",
            "items" : {
              "$ref" : "#/components/schemas/MessageReference"
            }
          },
          "count" : {
            "type" : "integer",
            "description" : "Count of all messages",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "description" : "limit number of references.",
            "format" : "int32"
          },
          "offset" : {
            "type" : "integer",
            "description" : "offset of references.",
            "format" : "int32"
          }
        },
        "description" : "A Message Reference object to be used when multiple messages could be retrieved from an API operation",
        "$id" : "https://joinup.ec.europa.eu/collection/api4dt/solution/.../message-reference-schema-list.json",
        "$schema" : "https://json-schema.org/draft/2020-12/schema"
      },
...
      "jws-compact-detached" : {
        "title" : "The message-level and payload signature",
        "pattern" : "^[A-Za-z0-9_-]+((?:\\.[A-Za-z0-9_-]+){2}|(?:(\\.\\.)[A-Za-z0-9_-]+){1})",
        "type" : "string",
        "description" : "The JWS compact representation schema defines the string pattern as a regular expression, denoting the structure a JWS compact token MUST follow to be a valid compact JWS Representation. It is used for defining the value of the message-level signature and of the payload signature",
        "format" : "jws-compact-detached"
      },
      "Problem" : {
        "title" : "A Problem Details object (RFC 7807) defined by the ISA² IPS REST API Core Profile",
        "required" : [ "status", "title", "type" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short summary of the problem type, written in English and readable for engineers (usually not suited for non technical stakeholders and not localized).",
            "example" : "Service Unavailable"
          },
          "type" : {
            "type" : "string",
            "description" : "An URI reference that identifies the problem type. When dereferenced, it SHOULD provide human-readable documentation for the problem type (e.g. using HTML)",
            "format" : "uri",
            "default" : "about:blank"
          },
          "status" : {
            "maximum" : 600,
            "exclusiveMaximum" : true,
            "minimum" : 200,
            "type" : "integer",
            "description" : "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format" : "int32",
            "example" : 503
          },
          "detail" : {
            "type" : "string",
            "description" : "A human-readable explanation specific to this occurrence of the problem",
            "format" : "string"
          },
          "instance" : {
            "type" : "string",
            "description" : "A URI reference that identifies the specific occurrence of the problem. It may or may not yieldfurther information if dereferenced.",
            "format" : "uri-reference"
          }
        },
        "description" : "A Problem Details object (RFC 7807) with ISA² IPS REST API extensions, used for signals (responses) to messages",
        "$id" : "https://joinup.ec.europa.eu/collection/api4dt/solution/.../messaging.schema.json",
        "$schema" : "https://json-schema.org/draft/2020-12/schema"
      },
...
      "SignalMessage" : {
        "title" : "A Signal message which extends Problem object (RFC 7807) defined by the ISA² IPS REST API Core Profile",
        "required" : [ "instance", "status", "title", "type" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short summary of the problem type, written in English and readable for engineers (usually not suited for non technical stakeholders and not localized).",
            "example" : "Service Unavailable"
          },
          "type" : {
            "type" : "string",
            "description" : "An URI reference that identifies the problem type. When dereferenced, it SHOULD provide human-readable documentation for the problem type (e.g. using HTML)",
            "format" : "uri",
            "default" : "about:blank"
          },
          "status" : {
            "maximum" : 600,
            "exclusiveMaximum" : true,
            "minimum" : 200,
            "type" : "integer",
            "description" : "The HTTP status code generated by the origin server for this occurrence of the ",
            "format" : "int32",
            "example" : 503
          },
          "detail" : {
            "type" : "string",
            "description" : "A human-readable explanation specific to this occurrence of the problem",
            "format" : "string"
          },
          "instance" : {
            "type" : "string",
            "description" : "A URI reference that identifies the specific occurrence of the  It may or may not yield further information if dereferenced.",
            "format" : "uri-reference"
          },
          "digest" : {
            "type" : "string",
            "description" : "The digest of the received message using the notation proposed in 'Digest Header' (https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-digest-headers)",
            "format" : "string",
            "example" : "sha-256=4REjxQ4yrqUVicfSKYNO/cF9zNj5ANbzgDZt3/h3Qxo="
          }
        },
        "description" : "A Signal message which extends Problem object (RFC 7807) defined by the ISA² IPS REST API Core Profile",
        "$id" : "https://joinup.ec.europa.eu/collection/api4dt/solution/.../messaging.schema.json",
        "$schema" : "https://json-schema.org/draft/2020-12/schema"
      }
    },
    "parameters" : {
      "messageId" : {
        "name" : "messageId",
        "in" : "path",
        "description" : "The MessageId is the unique identifier of the message submitted. It MUST be defined by the client. It is used for reliable messaging for guaranteeing the at-most-once message submission (no duplicate message-ids are allowed by the server implementing the API).",
        "required" : true,
        "schema" : {
          "type" : "string"
        },
        "example" : "dde12f67-c391-4851-8fa2-c07dd8532efd"
      },
...
      "Original-Sender" : {
        "name" : "Original-Sender",
        "in" : "header",
        "description" : "The Original Sender represents the authenticated entity acting as the user that sends the message using the client. Following the API Core Profile, the original sender MUST be represented with an OpenID Connect token, in the form of Compact JWT Token.",
        "schema" : {
          "$ref" : "#/components/schemas/jws-compact-detached"
        },
        "example" : "eyJhbGciOiJIUzINiIsnRI6IkpXVCJ9.eyJzdWIiOiIxM0NTY3ODkwIiwibmIiwiaWF0IoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMePOk6yJV_adQssw5c"
      },
      "Final-Recipient" : {
        "name" : "Final-Recipient",
        "in" : "header",
        "description" : "The Final Recipient is the entity for whom the message is sent. Can be a single identifier or a set of identifiers, as described in the Recipient Addressing Schemes section of the specification. The representation of the final recipient is out of scope of the current specification.",
        "schema" : {
          "type" : "string"
        },
        "example" : "9999::333222111"
      },
      "Timestamp" : {
        "name" : "Timestamp",
        "in" : "header",
        "description" : "The Timestamp is the exact date and time at which the message was sent. It is provided by the client and verified by the server.",
        "schema" : {
          "type" : "string",
          "format" : "date-time"
        },
        "example" : "2021-06-17T08:30:00Z"
      },
      "Edel-Message-Sig" : {
        "name" : "Edel-Message-Sig",
        "in" : "header",
        "description" : "The Edel-Message-Sig carries the signature of the HTTP Message following the API Core Profile on Message-Level Security. Following the light context constraints, the signature is optional for the client messages, but is RECOMMENDED for server messages.",
        "schema" : {
          "$ref" : "#/components/schemas/jws-compact-detached"
        },
        "example" : "eyJhbGciOiJIUzINiIsnRI6IkpXVCJ9.eyJzdWIiOiIxM0NTY3ODkwIiwibmIiwiaWF0IoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMePOk6yJV_adQssw5c"
      }
    },
    "headers" : {
      "Original-Sender" : {
        "description" : "The Original Sender represents the authenticated entity acting as the user that sends the message using the client. Following the API Core Profile, the original sender MUST be represented with an OpenID Connect token, in the form of Compact JWT Token.",
        "schema" : {
          "$ref" : "#/components/schemas/jws-compact-detached"
        }
      },
      "Final-Recipient" : {
        "description" : "The Final Recipient is the entity for whom the message is sent. Can be a single identifier or a set of identifiers, as described in the Recipient Addressing Schemes section of the specification. The representation of the final recipient is out of scope of the current specification.",
        "schema" : {
          "type" : "string"
        }
      },
      "Timestamp" : {
        "description" : "The Timestamp is the exact date and time at which the message was sent. It is provided by the client and verified by the server.",
        "schema" : {
          "type" : "string",
          "format" : "date-time"
        }
      },
      "Edel-Message-Sig" : {
        "description" : "The Edel-Message-Sig carries the signature of the HTTP Message following the API Core Profile on Message-Level Security. Following the light context constraints, the signature is optional for the client messages, but is RECOMMENDED for server messages.",
        "schema" : {
          "$ref" : "#/components/schemas/jws-compact-detached"
        }
      },
...
    },
    "securitySchemes" : {
      "DSD_ClientCredentials_OAuthSecurity" : {
        "type" : "oauth2",
        "flows" : {
          "clientCredentials" : {
            "tokenUrl" : "https://dev-24443841.okta.com/oauth2/aus1096gcr9r537Ut5d7/v1/token",
            "scopes" : {
              "ROLE_DSD" : "Authorization DSD token"
            }
          }
        }
      }
    }
  }
}